Commit №3 Deleting and Creating function of Fridge Fragment. Create the Image Adaptor

+++ app/src/main/java/com/example/dsamo/foodmanager/models/database/entity/FridgeListProducts.java
@@ -12,24 +12,24 @@ import static android.arch.persistence.room.ForeignKey.CASCADE;
         foreignKeys = {@ForeignKey(entity = Fridge.class, parentColumns = "id", childColumns = "fridge_id", onDelete =  CASCADE),
         @ForeignKey(entity = Product.class, parentColumns = "id", childColumns = "product_id", onDelete =  CASCADE)})
 public class FridgeListProducts {
-    private long fridge_id;
-    private long product_id;
+    private int fridge_id;
+    private int product_id;
 
     public FridgeListProducts(){}
 
-    public long getFridge_id() {
+    public int getFridge_id() {
         return fridge_id;
     }
 
-    public void setFridge_id(long fridge_id) {
+    public void setFridge_id(int fridge_id) {
         this.fridge_id = fridge_id;
     }
 
-    public long getProduct_id() {
+    public int getProduct_id() {
         return product_id;
     }
 
-    public void setProduct_id(long product_id) {
+    public void setProduct_id(int product_id) {
         this.product_id = product_id;
     }
 }
